backend/requirements.txt
fastapi[standard]==0.115.5
openai==1.55.2
playwright==1.18
pydantic-settings==2.6.1